Renamed actions & CFs, require updates for actions
Action names better reflect the change and not the ticket. Some actions
now require an update - for example, rejecting a change request - the
reviewer should have to tell the requestor *why* the change was
rejected.

Partially Deployed now has some actions as it is not necessarily an end
state. It could be the end, it could be marked failed, etc.