Make saved searches show in query builder Load menu
Saved searches need to have an attribute name of SavedSearch; the
description should be the actual name of the search, otherwise the Load
menu in the query builder doesn't know to include them in the list of
selections.

An upgrade script is needed to correct this for any installed instances
of the change management extension.